changes in .h to work with ofxEtherdream
